pyproject.toml
[tool.poetry]
name = "expense-tracker-api"
version = "0.1.0"
description = ""
authors = ["Metin100 <9455@holbertonstudents.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
sqlmodel = "^0.0.22"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
cryptography = "^44.0.0"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
